+++ src/briefy/choreographer/actions/leica/assignment.py
@@ -6,33 +6,37 @@ from zope.component import adapter
 from zope.interface import implementer
 
 
-@adapter(events.ILaureAssignmentValidated)
-@implementer(ILeicaAction)
-class AssignmentValidated(LeicaAction):
-    """Send validated assignment data to Leica."""
+class LeicaAssignment(LeicaAction):
+    """Handle assignment events to be sent to Leica."""
 
     entity = 'Assignment'
+
+
+@adapter(events.ILaureAssignmentValidated)
+@implementer(ILeicaAction)
+class AssignmentValidated(LeicaAssignment):
+    """Send validated assignment data to Leica."""
 
 
 @adapter(events.ILaureAssignmentRejected)
 @implementer(ILeicaAction)
-class AssignmentInvalidated(LeicaAction):
+class AssignmentInvalidated(LeicaAssignment):
     """Send invalidated assignment data to Leica."""
 
-    entity = 'Assignment'
+
+@adapter(events.ILaureAssignmentIgnored)
+@implementer(ILeicaAction)
+class AssignmentIgnored(LeicaAssignment):
+    """Send ignored assignment data to Leica."""
 
 
 @adapter(events.ILaureAssignmentCopied)
 @implementer(ILeicaAction)
-class AssignmentCopied(LeicaAction):
+class AssignmentCopied(LeicaAssignment):
     """Send data on assignment copied data to Leica."""
-
-    entity = 'Assignment'
 
 
 @adapter(events.ILaureAssignmentCopyFailure)
 @implementer(ILeicaAction)
-class AssignmentCopyFailure(LeicaAction):
+class AssignmentCopyFailure(LeicaAssignment):
     """Send notification that assets failed to be copied to Leica."""
-
-    entity = 'Assignment'